Move _Assets and Archive into Common namespace
