+++ OpenKh.Tests/Common/res/ps4archive.bin
